requirements.txt
python-dotenv
langchain
langchain-community
langchain-groq
ipykernel
streamlit